requirements.txt
aiohttp==2.0.7
cchardet==2.1.0
gunicorn==19.7.1
PyJWT==1.5.0